requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.2
python-dateutil>=2.8.2
lxml>=5.1.0
psycopg2-binary>=2.9.9
pandas>=2.1.0
python-dotenv>=1.0.0
